fix(awg): actually apply the obfuscation, and make keepalive a range
Both found by creating a real AmneziaWG inbound and looking at what the module
reported, which no test in the repo would have caught.

The obfuscation was never applied. Every layer worked -- the parameters
validated, encoded, and had a ConfigureParams to push them -- and nothing called
it. The panel brought up a device with the right port, the right peer and the
right allowed IPs, reported success, and served plain WireGuard: a tunnel every
client configured for AmneziaWG fails to use, with no error on either side. The
core now applies them after the device exists, from Reconcile and ApplyInstance.

Keepalive is a u16 RANGE, the third parameter of this shape to catch me out.
timers.c calls u16_range_pick_one on it, so the module draws a fresh interval
each time; encoded as a bare 25 it became [25, 0], which the tools print as
"25-0" and which asks the kernel to pick from an inverted range. The device now
reports "persistent keepalive: 25".

The lesson is the same one this whole session keeps teaching: a layer that
compiles, lints, unit-tests and reads correctly can still do nothing at all, and
the only thing that says so is running it against the real system.

+++ internal/awg/decode.go
@@ -151,7 +151,9 @@ func decodePeer(data []byte) (wgtypes.Peer, error) {
 		case peerEndpoint:
 			peer.Endpoint = decodeSockaddr(decoder.Bytes())
 		case peerPersistentKeepaliveInterval:
-			peer.PersistentKeepaliveInterval = time.Duration(decoder.Uint32()) * time.Second
+			// The low half is the range's lower bound, which is the interval asked
+			// for when both bounds match.
+			peer.PersistentKeepaliveInterval = time.Duration(uint16(decoder.Uint32())) * time.Second
 		case peerLastHandshakeTime:
 			peer.LastHandshakeTime = decodeTimespec(decoder.Bytes())
 		case peerRxBytes:
